Unpin manifest
This worked on other machines, with a caveat: pinning needs a connection
to specific guix commits, as versions quickly upgrade. I had to adjust
some package versions to match the guix commit of the target machine. So
for now I'm choosing to pin when I need snapshots of a profile saved for
broader consumption.