upgrade core emacs manifests
Exchanged emacs-matrix-client with emacs-ement, introduced in guix
commit 140b486437670ce95cb24a935b58cba52a9dac31 on 2021-11-10.

Removed heavy libraries like shellcheck (~2 GB!), as doom-emacs does not need
them to build and I was not using them frequently.

If I do need anything that was removed in this commit, I can design
workflows which use env, guix shell, and guix home to add these
dependencies, in docker containers or vms.